Update infer_omgsr_f.py
Traceback (most recent call last):
  File "D:\OMGSR\infer\infer_omgsr_f.py", line 156, in <module>
    main(args)
  File "D:\OMGSR\infer\infer_omgsr_f.py", line 101, in main
    output_pil = transforms.ToPILImage()(output_image[0].cpu())
  File "C:\Users\LHKJACKY\.conda\envs\OMGSR2\lib\site-packages\torchvision\transforms\transforms.py", line 234, in __call__
    return F.to_pil_image(pic, self.mode)
  File "C:\Users\LHKJACKY\.conda\envs\OMGSR2\lib\site-packages\torchvision\transforms\functional.py", line 266, in to_pil_image
    pic = pic.numpy(force=True)
TypeError: Got unsupported ScalarType BFloat16

+++ infer/infer_omgsr_f.py
@@ -98,7 +98,8 @@ def main(args):
 
         output_image = output_image * 0.5 + 0.5
         output_image = torch.clip(output_image, 0, 1)
-        output_pil = transforms.ToPILImage()(output_image[0].cpu())
+        #output_pil = transforms.ToPILImage()(output_image[0].cpu())
+        output_pil = transforms.ToPILImage()(output_image[0].cpu().float())  # Converts bfloat16 -> float32
 
         if args.align_method == 'adain':
             output_pil = adain_color_fix(target=output_pil, source=input_image)
